Implemented classes for frontend services
