+++ Frontend/src/components/Todos/TaskList.tsx
@@ -1,11 +1,7 @@
 import React from "react";
 import { ListGroupStyled, ListItemStyled, CheckboxStyled } from "../../styles/TodoStyles";
 import Button from "../common/Button/Button";
-import {
-  handleDeleteTask,
-  handleEditTask,
-  handleToggleCompleted,
-} from "../../services/todoServices";
+import TodoServices from "../../services/TodoServices";
 import { TaskListProps } from "../../interface/Todo/index";
 import { EDIT_LABEL, DELETE_LABEL } from "../../constants/LABELS";
 
@@ -26,7 +22,7 @@ const TaskList: React.FC<TaskListProps> = ({
             <CheckboxStyled
               type="checkbox"
               checked={todo.completed}
-              onChange={() => handleToggleCompleted(todo.id, setTodos, todos)}
+              onChange={() => TodoServices.handleToggleCompleted(todo.id, setTodos, todos)}
             />
             {todo.task}
           </div>
@@ -34,14 +30,14 @@ const TaskList: React.FC<TaskListProps> = ({
             <Button
               buttonStyle="warning"
               width="auto"
-              onClick={() => handleEditTask(todo, setEditingTask, setNewTask)}
+              onClick={() => TodoServices.handleEditTask(todo, setEditingTask, setNewTask)}
             >
               {EDIT_LABEL}
             </Button>
             <Button
               buttonStyle="danger"
               width="auto"
-              onClick={() => handleDeleteTask(todo.id, setTodos, todos)}
+              onClick={() => TodoServices.handleDeleteTask(todo.id, setTodos, todos)}
             >
               {DELETE_LABEL}
             </Button>
