--- Frontend/src/interface/Todo.ts
@@ -1,5 +1,0 @@
-export interface Todo {
-  id: number;
-  task: string;
-  completed: boolean;
-}
