--- Frontend/src/services/authServices/index.ts
@@ -1,3 +1,0 @@
-export { default as handleLogin } from "./handleLogin";
-export { default as handleLogout } from "./handleLogout";
-export { default as handleRegister } from "./handleRegister";
