--- Frontend/src/services/todoServices/index.ts
@@ -1,6 +1,0 @@
-export { default as handleAddTask } from "./handleAddTask";
-export { default as handleDeleteTask } from "./handleDeleteTask";
-export { default as handleEditTask } from "./handleEditTask";
-export { default as handleToggleCompleted } from "./handleToggleCompleted";
-export { default as handleUpdateTask } from "./handleUpdateTask";
-export { default as fetchTodos } from "./fetchTodos";
